player: validate stream options and reuse go-ffmpeg-ffi typed fields
Review follow-up. RTSPTransport is validated against FFmpeg's transport
names at the public API, so a misconfigured value fails fast with an
error naming the field instead of an opaque EINVAL retried forever by
reconnect loops. The URL-scheme gate is gone: FFmpeg ignores the option
for demuxers it does not apply to, and the gate silently dropped it for
satip:// and SDP sources the RTSP demuxer does handle.

ProbeSize and AnalyzeDuration now use go-ffmpeg-ffi's typed
DecoderOptions fields, which own the FFmpeg key names and take
precedence over raw AVOptions; they apply to any open, not only live
ones, and ProbeSize is raised to FFmpeg's minimum of 32 bytes instead
of failing the open. The timeout documentation now states the
per-protocol reality: ConnTimeout bounds RTSP TCP socket I/O, rw_timeout
bounds AVIO-based protocols such as HTTP, and RTSP over UDP transport
is unbounded.

+++ backend.go
@@ -35,7 +35,7 @@ type backendOpenOptions struct {
 	ConnTimeout      time.Duration
 	ReadTimeout      time.Duration
 	RTSPTransport    string
-	ProbeSize        int64
+	ProbeSize        int
 	AnalyzeDuration  time.Duration
 }
